🌱 Extract useStandaloneOrbit hook from StandaloneOrbitDialog
StandaloneOrbitDialog.tsx drops from 384 to 275 lines and from 15 to 2
hook calls by moving all state, cluster selection and orbit-mission
persistence logic into useStandaloneOrbit.ts. `clusters` is now memoized
so the pre-existing exhaustive-deps warning is fixed rather than moved.

Pure refactor — no behaviour change.

Refs #21787
